Fix app crashing if built with GCC >= 13.1
App worked fine in runtime 22.08 (GCC 12.2.0)
but starting crashing with runtime 23.08 which
comes with GCC 13.1.0

After debugging the crash, the problem was the
order in which the class member initializers were
executed (more info here[1]) because the code
had some classes initializing Gtk::RadioButton
members with a reference to a Gtk::RadioButtonGroup
that was still not initialized, because that
Gtk::RadioButtonGroup member was declared in the
header file after the Gtk::RadioButton in question.

Apparently GCC 12 was more resilient to this kind
of inconsistency.

[1] https://stackoverflow.com/a/1828058